git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/ltl-model-checking@31517 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/statistics/StatisticsUtils.java
@@ -43,7 +43,7 @@ public class StatisticsUtils {
     return String.format("%1.0f%%", val / full * 100);
   }
 
-  public static String valueWithPercentage(int value, int totalCount) {
+  public static String valueWithPercentage(long value, int totalCount) {
     return value + " (" + toPercent(value, totalCount) + ")";
   }
